# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mood to Music API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  historyId: string;
  songId: string;
  songName: string;
  artist: string;
  moodId: string;
  moodName: string;
  emoji?: string;
  playedDate: string;
  time: string;
}