# lib/api-zod/src/generated/types/moodSelection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mood to Music API
 * OpenAPI spec version: 0.1.0
 */

export interface MoodSelection {
  selectionId: string;
  moodId: string;
  moodName: string;
  emoji?: string;
  date: string;
  time: string;
}