# lib/api-zod/src/generated/types/playlistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mood to Music API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface PlaylistDetail {
  playlistId: string;
  playlistName: string;
  createdDate: string;
  songs: Song[];
}